Renamed Tracker to TrackerAdapter
