+++ src/interfaces/i-tracker-adapter.js
